Attendance Initial back end

+++ server/src/main/java/com/postgresql/pgms/model/program.java
@@ -9,7 +9,7 @@ import lombok.Data;
 public class program {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private long programID;
+    private Integer programID;
 
     private String name;
